+++ com.revolsys.open.core/src/main/java/com/revolsys/gis/model/coordinates/list/ReverseCoordinatesList.java
@@ -20,7 +20,7 @@ public class ReverseCoordinatesList extends AbstractCoordinatesList {
   }
 
   @Override
-  public byte getNumAxis() {
+  public int getNumAxis() {
     return coordinateList.getNumAxis();
   }
