+++ com.revolsys.open.core/src/main/java/com/revolsys/gis/tin/Triangle.java
@@ -178,7 +178,7 @@ public class Triangle extends AbstractCoordinatesList {
   }
 
   @Override
-  public byte getNumAxis() {
+  public int getNumAxis() {
     return 3;
   }
