+++ com.revolsys.open.core/src/main/java/com/revolsys/jts/geom/impl/GeometryImpl.java
@@ -1429,7 +1429,7 @@ public abstract class GeometryImpl implements Geometry {
   }
 
   @Override
-  public byte getNumAxis() {
+  public int getNumAxis() {
     return (byte)geometryFactory.getNumAxis();
   }
